+++ src/Shadow.ts
@@ -2,7 +2,7 @@ import { Color } from './color/Color';
 import { config } from './config';
 import { Point } from './Point';
 import type { FabricObject } from './shapes/Object/FabricObject';
-import { TClassProperties } from './typedefs';
+import type { TClassProperties } from './typedefs';
 import { uid } from './util/internals/uid';
 import { pickBy } from './util/misc/pick';
 import { degreesToRadians } from './util/misc/radiansDegreesConversion';