+++ src/gradient/Gradient.ts
@@ -4,7 +4,7 @@ import { iMatrix } from '../constants';
 import { parseTransformAttribute } from '../parser/parseTransformAttribute';
 import type { FabricObject } from '../shapes/Object/FabricObject';
 import { FabricObject as BaseFabricObject } from '../shapes/Object/Object';
-import { TMat2D } from '../typedefs';
+import type { TMat2D } from '../typedefs';
 import { uid } from '../util/internals/uid';
 import { pick } from '../util/misc/pick';
 import { matrixToSVG } from '../util/misc/svgParsing';
@@ -15,7 +15,7 @@ import {
   parseGradientUnits,
   parseType,
 } from './parser';
-import {
+import type {
   ColorStop,
   GradientCoords,
   GradientOptions,