+++ src/parser/parseTransformAttribute.ts
@@ -1,6 +1,6 @@
 import { iMatrix } from '../constants';
 import { reNum } from './constants';
-import { TMat2D } from '../typedefs';
+import type { TMat2D } from '../typedefs';
 import { cleanupSvgAttribute } from '../util/internals/cleanupSvgAttribute';
 import {
   createRotateMatrix,