+++ src/shapes/Object/types/index.ts
@@ -1,7 +1,7 @@
-import { FabricObjectProps } from './FabricObjectProps';
+import type { FabricObjectProps } from './FabricObjectProps';
 
-export { SerializedObjectProps } from './SerializedObjectProps';
-export { FabricObjectProps };
+export type { SerializedObjectProps } from './SerializedObjectProps';
+export type { FabricObjectProps };
 
 export type TProps<T> = Partial<T> & Record<string, any>;